refactor: create @pnpm/fetching-types
