+++ packages/client/tsconfig.json
@@ -16,6 +16,9 @@
       "path": "../fetch"
     },
     {
+      "path": "../fetching-types"
+    },
+    {
       "path": "../git-fetcher"
     },
     {
