+++ packages/fetch/package.json
@@ -34,6 +34,7 @@
   "homepage": "https://github.com/pnpm/pnpm/blob/master/packages/fetch#readme",
   "dependencies": {
     "@pnpm/core-loggers": "workspace:^4.1.2",
+    "@pnpm/fetching-types": "workspace:^0.0.0",
     "@pnpm/npm-registry-agent": "workspace:^2.0.3",
     "@zkochan/retry": "^0.2.0",
     "node-fetch": "2.6.0",
