+++ packages/fetch/tsconfig.json
@@ -13,6 +13,9 @@
       "path": "../core-loggers"
     },
     {
+      "path": "../fetching-types"
+    },
+    {
       "path": "../npm-registry-agent"
     }
   ]
