+++ packages/fetching-types/README.md
@@ -1,0 +1,7 @@
+# @pnpm/fetching-types
+
+> Types for the fetching
+
+## License
+
+MIT
