+++ packages/fetching-types/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "@pnpm/fetching-types",
+  "version": "0.0.0",
+  "description": "Types for fetching",
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "engines": {
+    "node": ">=10.13"
+  },
+  "files": [
+    "lib",
+    "!*.map"
+  ],
+  "repository": "https://github.com/pnpm/pnpm/blob/master/packages/fetching-types",
+  "keywords": [
+    "pnpm",
+    "types"
+  ],
+  "author": "Zoltan Kochan <z@kochan.io> (https://www.kochan.io/)",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/pnpm/pnpm/issues"
+  },
+  "homepage": "https://github.com/pnpm/pnpm/blob/master/packages/fetching-types#readme",
+  "scripts": {
+    "compile": "rimraf lib tsconfig.tsbuildinfo && tsc --build",
+    "prepublishOnly": "pnpm run compile",
+    "test": "pnpm run compile"
+  },
+  "funding": "https://opencollective.com/pnpm",
+  "dependencies": {
+    "@types/node-fetch": "^2.5.7",
+    "@zkochan/retry": "^0.2.0"
+  }
+}
