+++ packages/npm-resolver/package.json
@@ -35,7 +35,7 @@
   },
   "dependencies": {
     "@pnpm/error": "workspace:1.2.0",
-    "@pnpm/fetch": "workspace:^2.0.2",
+    "@pnpm/fetching-types": "workspace:^0.0.0",
     "@pnpm/resolve-workspace-range": "workspace:1.0.1",
     "@pnpm/resolver-base": "workspace:7.0.3",
     "@pnpm/types": "workspace:6.2.0",
@@ -53,6 +53,7 @@
     "version-selector-type": "^3.0.0"
   },
   "devDependencies": {
+    "@pnpm/fetch": "workspace:^2.0.2",
     "@pnpm/logger": "3.2.2",
     "@pnpm/npm-resolver": "link:",
     "@types/mz": "^2.7.1",
