+++ packages/tarball-fetcher/package.json
@@ -36,17 +36,16 @@
   "dependencies": {
     "@pnpm/core-loggers": "workspace:^4.1.2",
     "@pnpm/error": "workspace:1.2.0",
-    "@pnpm/fetch": "workspace:^2.0.2",
     "@pnpm/fetcher-base": "workspace:8.0.2",
+    "@pnpm/fetching-types": "workspace:^0.0.0",
     "@zkochan/retry": "^0.2.0",
     "graceful-fs": "^4.2.4",
     "mz": "^2.7.0",
-    "path-temp": "^2.0.0",
-    "rimraf": "^3.0.2",
     "ssri": "6.0.1"
   },
   "devDependencies": {
     "@pnpm/cafs": "workspace:1.0.5",
+    "@pnpm/fetch": "workspace:^2.0.2",
     "@pnpm/logger": "3.2.2",
     "@pnpm/tarball-fetcher": "link:",
     "@types/graceful-fs": "^4.1.3",
